SLVS-2578 Move class into ReportView folder, so it can be reused
